+++ example_app/config/config.exs
@@ -9,3 +9,10 @@ config :example_app, ExampleApp.Repo,
 
 config :example_app,
   ecto_repos: [ExampleApp.Repo]
+
+config :example_app, :grid_codec,
+  schemas: %{
+    100 => "events",
+    99 => "bench_trace",
+    200 => "bench_sizes"
+  }